Move airbrake integration and SIGINT/SIGTERM interception to recurring-payments-processor, along with a new export of execute that will catch errors and call console.error with them, so that the airbrake routing through Errbit and to Slack works
